requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
asyncpg==0.27.0
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
kiwisolver==1.4.4
magic-filter==1.0.9
matplotlib==3.7.0
multidict==6.0.4
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
psycopg2==2.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
SQLAlchemy==2.0.4
typing_extensions==4.5.0
yarl==1.8.2
